+++ src/pybamm/models/full_battery_models/lithium_ion/base_lithium_ion_model.py
@@ -270,12 +270,10 @@ class BaseModel(pybamm.BaseBatteryModel):
                     ocp_model = ocp_submodels.SingleOpenCircuitPotential
                 elif ocp_option == "current sigmoid":
                     ocp_model = ocp_submodels.CurrentSigmoidOpenCircuitPotential
-                elif ocp_option == "Wycisk":
-                    pybamm.citations.register("Wycisk2022")
-                    ocp_model = ocp_submodels.WyciskOpenCircuitPotential
-                elif ocp_option == "Axen":
-                    pybamm.citations.register("Axen2022")
-                    ocp_model = ocp_submodels.AxenOpenCircuitPotential
+                elif ocp_option == "one-state hysteresis":
+                    ocp_model = ocp_submodels.OneStateHysteresisOpenCircuitPotential
+                elif ocp_option == "one-state differential capacity hysteresis":
+                    ocp_model = ocp_submodels.OneStateDifferentialCapacityHysteresisOpenCircuitPotential
                 elif ocp_option == "MSMR":
                     ocp_model = ocp_submodels.MSMROpenCircuitPotential
                 self.submodels[f"{domain} {phase} open-circuit potential"] = ocp_model(