+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 
 - Removed the IREE code from the IDAKLU solver ([#5080](https://github.com/pybamm-team/PyBaMM/pull/5080))
 - Removed support for Python 3.9 ([#5052](https://github.com/pybamm-team/PyBaMM/pull/5052))
+- Renamed the open-circuit potential models to be more descriptive. The options "Axen" and "Wycisk" are now "one-state hysteresis" and "one-state differential capacity hysteresis". The old classes are still available as aliases. ([#4893](https://github.com/pybamm-team/PyBaMM/pull/4893))
 - In OCP hysteresis models, users need to explicitly give the equilibrium, delithiation, and lithiation OCPs when using a hysteresis model. E.g., you must provide all three of "Negative electrode OCP [V]", "Negative electrode delithiation OCP [V]", and "Negative electrode lithiation OCP [V]". ([#4893](https://github.com/pybamm-team/PyBaMM/pull/4893))
 - Replaces `1-sign(i_vol)h` with `(1-sign(i_vol)h)/2` so that the parameter $\gamma$ reflects the natural rate of change. For backwards compatibility, parameters named "...hysteresis decay rate" should be multiplied by 2 to give the same results as before. ([#4893](https://github.com/pybamm-team/PyBaMM/pull/4893))
 # [v25.6.0](https://github.com/pybamm-team/PyBaMM/tree/v25.6.0) - 2025-05-27